Refactor into several files
